Ensure all nested routes are protected based on memberships and session. Rename components for clarity.
