+++ app/[slug]/(base)/page.tsx
@@ -1,4 +1,3 @@
-import {redirect} from 'next/navigation'
 import {Suspense} from 'react'
 import {ProjectsList} from '~/components/dashboard/Projects/ProjectsList'
 import prisma from '~/prisma'
@@ -6,7 +5,6 @@ import {getCurrentUser} from '~/utils/session'
 
 export default async function Dashboard({params}: {params: {slug: string}}) {
 	const user = await getCurrentUser()
-	if (!user) redirect('/')
 
 	const memberships = await prisma.membership.findFirst({
 		where: {
