+++ app/[slug]/(base)/settings/general/page.tsx
@@ -1,0 +1,7 @@
+export default function General({params}: {params: {slug: string}}) {
+	return (
+		<div className='flex flex-col gap-6 py-0.5'>
+			<h3>General</h3>
+		</div>
+	)
+}
