+++ app/[slug]/layout.tsx
@@ -1,48 +1,28 @@
-import DashboardHeader from '~/components/dashboard/Navigation/DashboardHeader'
+import {notFound, redirect} from 'next/navigation'
+import DashboardNav from '~/components/dashboard/Navigation/DashboardNavServer'
 import prisma from '~/prisma'
 import {getCurrentUser} from '~/utils/session'
 
-async function Teams() {
-	const user = await getCurrentUser()
-	if (!user) return <></>
-
-	const teams = await prisma.membership
-		.findMany({
-			where: {userId: user.id},
-			select: {
-				team: {
-					include: {
-						Project: {
-							select: {
-								id: true,
-								name: true
-							}
-						}
-					}
-				}
-			}
-		})
-		.then(memberships => memberships.map(m => m.team))
-
-	const projects = teams.reduce((projs, {Project}) => [...projs, ...Project], [])
-
-	return (
-		<DashboardHeader
-			user={user}
-			teams={teams}
-			projects={projects}
-		/>
-	)
-}
-
 export default async function RootLayout({
-	children
+	children,
+	params
 }: {
 	children: React.ReactNode
+	params: {slug: string}
 }) {
+	// Ensure all dashboard routes are protected
+	const user = await getCurrentUser()
+	if (!user) return redirect('/')
+
+	// Ensure user is a member of the team being accessed
+	const team = await prisma.team.findFirst({
+		where: {slug: params.slug, memberships: {some: {userId: user.id}}}
+	})
+	if (!team) return notFound()
+
 	return (
 		<div className='relative min-h-screen w-full px-8'>
-			<Teams />
+			<DashboardNav user={user} />
 			<div className='w-full'>{children}</div>
 		</div>
 	)
