+++ app/[slug]/project/[projectId]/layout.tsx
@@ -1,6 +1,3 @@
-import {getServerSession} from 'next-auth'
-import {redirect} from 'next/navigation'
-import {authOptions} from '~/authOptions'
 import ProjectNav from '~/components/dashboard/Navigation/ProjectNav'
 
 export default async function Layout({
@@ -10,9 +7,6 @@ export default async function Layout({
 	params: {slug: string; projectId: string}
 	children: React.ReactNode
 }) {
-	const session = await getServerSession(authOptions)
-	if (!session) redirect('/')
-
 	return (
 		<div className='space-y-4'>
 			<ProjectNav
